## Description

Fixes #1132 

## Checklist:

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [x] ~I have added appropriate tests that cover the new/changed
functionality.~
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes, or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

+++ src/qdmi/na/App.cpp
@@ -400,8 +400,8 @@ int main(int argc, char* argv[]) {
   // `main` functions should not throw exceptions. Apparently, the
   // initialization of a vector can throw exceptions, so we catch them here.
   try {
-    argVec.reserve(argc);
-    for (const auto& arg : std::span(argv, argc)) {
+    argVec.reserve(static_cast<size_t>(argc));
+    for (const auto& arg : std::span(argv, static_cast<size_t>(argc))) {
       argVec.emplace_back(arg);
     }
   } catch (std::exception& e) {
